librlist: fix unsigned difference comparison to zero
Problem: CodeQL scanning detected an unsigned comparison to zero
in librlist errlist_append().

Fix the function to cast the unsigned size_t value to int before
the difference so that the comparison is made with a signed result.

+++ src/common/librlist/rlist.c
@@ -678,7 +678,7 @@ static int errlist_concat (zlistx_t *l, char *buf, size_t len)
 
     s = zlistx_first (l);
     while (s) {
-        if (len - n > 0)
+        if ((int)len - n > 0)
             strncpy (buf + n, s, len - n);
         n += strlen (s);
         s = zlistx_next (l);
